+++ lib/graphs.dart
@@ -3,5 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 export 'src/crawl_async.dart' show crawlAsync;
+export 'src/shortest_path.dart' show shortestPath, shortestPaths;
 export 'src/strongly_connected_components.dart'
     show stronglyConnectedComponents;